chore(api): move participant repartition model from shared to prescription context
