+++ api/src/prescription/organization-learner/domain/models/ParticipantRepartition.js
